+++ pypsa/statistics.py
@@ -14,9 +14,11 @@ __copyright__ = (
 
 import logging
 from functools import reduce, wraps
+from inspect import signature
 
 import numpy as np
 import pandas as pd
+from deprecation import deprecated
 
 from pypsa.descriptors import nominal_attrs
 
@@ -41,11 +43,20 @@ def get_bus_and_carrier(n, c, port="", nice_names=True):
     """
     Get the buses and nice carrier names for a component.
     """
-    if port == "":
-        bus = "bus0" if "bus" not in n.df(c) else "bus"
-    else:
-        bus = f"bus{port}"
+    bus = f"bus{port}"
     return [n.df(c)[bus].rename("bus"), get_carrier(n, c, nice_names=nice_names)]
+
+
+def get_bus_unit_and_carrier(n, c, port="", nice_names=True):
+    """
+    Get the buses and nice carrier names for a component.
+    """
+    bus = f"bus{port}"
+    return [
+        n.df(c)[bus].rename("bus"),
+        n.df(c)[bus].map(n.buses.unit).rename("unit"),
+        get_carrier(n, c, nice_names=nice_names),
+    ]
 
 
 def get_name_bus_and_carrier(n, c, port="", nice_names=True):
@@ -112,6 +123,8 @@ def get_weightings(n, c):
         return n.snapshot_weightings["objective"]
 
 
+# TODO: remove in 0.29
+@deprecated(deprecated_in="0.28", removed_in="0.29")
 def get_ports(n, c):
     """
     Get a list of existent ports of a component.
@@ -119,26 +132,7 @@ def get_ports(n, c):
     return [col[3:] for col in n.df(c) if col.startswith("bus")]
 
 
-def port_mask(n, c, port="", bus_carrier=None):
-    """
-    Get a mask of components which are optionally connected to a bus with a
-    given carrier.
-    """
-    if bus_carrier is None:
-        mask = n.df(c)[f"bus{port}"] != ""
-    elif isinstance(bus_carrier, str):
-        mask = n.df(c)[f"bus{port}"].map(n.buses.carrier).eq(bus_carrier)
-    elif isinstance(bus_carrier, list):
-        mask = n.df(c)[f"bus{port}"].map(n.buses.carrier).isin(bus_carrier)
-    else:
-        raise ValueError(
-            f"Argument `bus_carrier` must be a string or list, got {type(bus_carrier)}"
-        )
-    return mask.astype(int)
-
-
-def port_efficiency(n, c, port="", bus_carrier=None):
-    mask = port_mask(n, c, port=port, bus_carrier=bus_carrier)
+def port_efficiency(n, c, port=""):
     if (port == "") or "efficiency" not in n.df(c):
         efficiency = 1
     elif port == "0":
@@ -147,7 +141,7 @@ def port_efficiency(n, c, port="", bus_carrier=None):
         efficiency = n.df(c)["efficiency"]
     else:
         efficiency = n.df(c)[f"efficiency{port}"]
-    return mask * efficiency
+    return efficiency
 
 
 def get_transmission_branches(n, bus_carrier=None):
@@ -190,6 +184,25 @@ def get_transmission_carriers(n, bus_carrier=None):
     )
 
 
+def get_grouping(n, c, groupby, port=None, nice_names=False) -> [pd.Series, list]:
+    by = None
+    level = None
+    if callable(groupby):
+        if "port" in signature(groupby).parameters:
+            by = groupby(n, c, port=port, nice_names=nice_names)
+        else:
+            by = groupby(n, c, nice_names=nice_names)
+    elif isinstance(groupby, list):
+        by = [n.df(c)[key] for key in groupby]
+    elif isinstance(groupby, str):
+        by = n.df(c)[groupby]
+    elif groupby is not False:
+        ValueError(
+            f"Argument `groupby` must be a function, list, string, False or dict, got {type(groupby)}"
+        )
+    return dict(by=by, level=level)
+
+
 def aggregate_timeseries(df, weights, agg="sum"):
     "Calculate the weighted sum or average of a DataFrame or Series."
     if isinstance(df.index, pd.MultiIndex):
@@ -210,19 +223,55 @@ def aggregate_timeseries(df, weights, agg="sum"):
     return df.agg(agg)
 
 
+def filter_active_assets(n, c, df: [pd.Series, pd.DataFrame]):
+    """
+    For static values iterate over periods and concat values.
+    """
+    if not isinstance(n.snapshots, pd.MultiIndex) or isinstance(df, pd.DataFrame):
+        return df
+    per_period = {}
+    for p in n.investment_periods:
+        per_period[p] = df[n.get_active_assets(c, p).loc[df.index]]
+    return pd.concat(per_period, axis=1)
+
+
+def filter_bus_carrier(n, c, port, bus_carrier, df):
+    """
+    Filter the DataFrame for components which are connected to a bus with
+    carrier `bus_carrier`.
+    """
+    if bus_carrier is None:
+        return df
+
+    ports = n.df(c).loc[df.index, f"bus{port}"]
+    port_carriers = ports.map(n.buses.carrier)
+    if isinstance(bus_carrier, str):
+        if bus_carrier in n.buses.carrier.unique():
+            return df[port_carriers == bus_carrier]
+        else:
+            return df[bus_carrier in port_carriers]
+    elif isinstance(bus_carrier, list):
+        return df[port_carriers.isin(bus_carrier)]
+    else:
+        raise ValueError(
+            f"Argument `bus_carrier` must be a string or list, got {type(bus_carrier)}"
+        )
+
+
 def aggregate_components(
     n,
     func,
     agg="sum",
     comps=None,
     groupby=None,
+    at_port=None,
+    bus_carrier=None,
     nice_names=True,
 ):
     """
     Apply a function and group the result for a collection of components.
     """
     d = {}
-    kwargs = {}
 
     if is_one_component := isinstance(comps, str):
         comps = [comps]
@@ -234,38 +283,33 @@ def aggregate_components(
         if n.df(c).empty:
             continue
 
-        if callable(groupby):
-            grouping = groupby(n, c, nice_names=nice_names)
-        elif isinstance(groupby, list):
-            grouping = [n.df(c)[key] for key in groupby]
-        elif isinstance(groupby, str):
-            grouping = n.df(c)[groupby]
-        elif isinstance(groupby, dict):
-            grouping = None
-            kwargs = groupby
-        elif groupby is not False:
-            ValueError(
-                f"Argument `groupby` must be a function, list, string, False or dict, got {type(groupby)}"
-            )
+        ports = [col[3:] for col in n.df(c) if col.startswith("bus")]
+        if not at_port:
+            ports = [ports[0]]
 
-        df = func(n, c)
-        if isinstance(n.snapshots, pd.MultiIndex) and not isinstance(df, pd.DataFrame):
-            # for static values we have to iterate over periods and concat
-            per_period = {}
-            for p in n.investment_periods:
-                per_period[p] = df[n.get_active_assets(c, p).loc[df.index]]
-            df = pd.concat(per_period, axis=1)
-        d[c] = (
-            df.rename_axis("name")
-            if groupby is False
-            else df.groupby(grouping, **kwargs).agg(agg)
-        )
+        df = []
+        for port in ports:
+            vals = func(n, c, port)
+            vals = filter_active_assets(n, c, vals)  # for multiinvest
+            vals = filter_bus_carrier(n, c, port, bus_carrier, vals)
+
+            # unit tracker
+            if groupby is not False:
+                grouping = get_grouping(n, c, groupby, port=port, nice_names=nice_names)
+                vals = vals.groupby(**grouping).agg(agg)
+            df.append(vals)
+
+        df = pd.concat(df, copy=False) if len(df) > 1 else df[0]
+        if not df.index.is_unique:
+            df = df.groupby(level=df.index.names).agg(agg)
+        d[c] = df
 
     if d == {}:
         return pd.Series([])
     if is_one_component:
         return d[c]
-    return pd.concat(d, names=["component", *d[list(d.keys())[0]].index.names])
+    index_names = ["component"] + df.index.names
+    return pd.concat(d, names=index_names)[lambda ds: ds != 0]
 
 
 def pass_empty_series_if_keyerror(func):
@@ -292,6 +336,7 @@ class Groupers:
     get_bus_and_carrier_and_bus_carrier = staticmethod(
         get_bus_and_carrier_and_bus_carrier
     )
+    get_bus_unit_and_carrier = staticmethod(get_bus_unit_and_carrier)
 
 
 class StatisticsAccessor:
@@ -345,14 +390,16 @@ class StatisticsAccessor:
             )
             kwargs.pop("aggregate_time")
 
+        # TODO replace dispatch by energy_balance
+
         funcs = [
             self.optimal_capacity,
             self.installed_capacity,
-            self.capacity_factor,
+            self.supply,
+            self.withdrawal,
             self.dispatch,
             self.transmission,
-            self.withdrawal,
-            self.supply,
+            self.capacity_factor,
             self.curtailment,
             self.capex,
             self.opex,
@@ -379,6 +426,7 @@ class StatisticsAccessor:
         comps=None,
         aggregate_groups="sum",
         groupby=None,
+        at_port=False,
         bus_carrier=None,
         nice_names=True,
     ):
@@ -394,15 +442,8 @@ class StatisticsAccessor:
         n = self._parent
 
         @pass_empty_series_if_keyerror
-        def func(n, c):
+        def func(n, c, port):
             col = n.df(c).eval(f"{nominal_attrs[c]}_opt * capital_cost")
-            if bus_carrier is not None:
-                masks = [
-                    port_mask(n, c, port=port, bus_carrier=bus_carrier)
-                    for port in get_ports(n, c)
-                ]
-                mask = reduce(np.logical_or, masks)
-                col = col[mask.astype(bool)]
             return col
 
         df = aggregate_components(
@@ -411,6 +452,8 @@ class StatisticsAccessor:
             comps=comps,
             agg=aggregate_groups,
             groupby=groupby,
+            at_port=at_port,
+            bus_carrier=bus_carrier,
             nice_names=nice_names,
         )
         df.attrs["name"] = "Capital Expenditure"
@@ -422,6 +465,7 @@ class StatisticsAccessor:
         comps=None,
         aggregate_groups="sum",
         groupby=None,
+        at_port=False,
         bus_carrier=None,
         nice_names=True,
     ):
@@ -429,24 +473,14 @@ class StatisticsAccessor:
         Calculate the capital expenditure of already built components of the
         network in given currency.
 
-        If `bus_carrier` is given, the capacity is weighed by the output efficiency
-        of components at buses with carrier `bus_carrier`.
-
         For information on the list of arguments, see the docs in
         `Network.statistics` or `pypsa.statistics.StatisticsAccessor`.
         """
         n = self._parent
 
         @pass_empty_series_if_keyerror
-        def func(n, c):
+        def func(n, c, port):
             col = n.df(c).eval(f"{nominal_attrs[c]} * capital_cost")
-            if bus_carrier is not None:
-                masks = [
-                    port_mask(n, c, port=port, bus_carrier=bus_carrier)
-                    for port in get_ports(n, c)
-                ]
-                mask = reduce(np.logical_or, masks)
-                col = col[mask.astype(bool)]
             return col
 
         df = aggregate_components(
@@ -455,9 +489,49 @@ class StatisticsAccessor:
             comps=comps,
             agg=aggregate_groups,
             groupby=groupby,
+            at_port=at_port,
+            bus_carrier=bus_carrier,
             nice_names=nice_names,
         )
         df.attrs["name"] = "Capital Expenditure Fixed"
+        df.attrs["unit"] = "currency"
+        return df
+
+    def expanded_capex(
+        self,
+        comps=None,
+        aggregate_groups="sum",
+        groupby=None,
+        at_port=False,
+        bus_carrier=None,
+        nice_names=True,
+    ):
+        """
+        Calculate the capex of expanded capacities of the network components in
+        currency.
+
+        For information on the list of arguments, see the docs in
+        `Network.statistics` or `pypsa.statistics.StatisticsAccessor`.
+        """
+        df = self.capex(
+            comps=comps,
+            aggregate_groups=aggregate_groups,
+            groupby=groupby,
+            at_port=at_port,
+            bus_carrier=bus_carrier,
+            nice_names=nice_names,
+        ).sub(
+            self.installed_capex(
+                comps=comps,
+                aggregate_groups=aggregate_groups,
+                groupby=groupby,
+                at_port=at_port,
+                bus_carrier=bus_carrier,
+                nice_names=nice_names,
+            ),
+            fill_value=0,
+        )
+        df.attrs["name"] = "Capital Expenditure Expanded"
         df.attrs["unit"] = "currency"
         return df
 
@@ -466,6 +540,7 @@ class StatisticsAccessor:
         comps=None,
         aggregate_groups="sum",
         groupby=None,
+        at_port=False,
         bus_carrier=None,
         storage=False,
         nice_names=True,
@@ -488,22 +563,12 @@ class StatisticsAccessor:
             comps = ("Store", "StorageUnit")
 
         @pass_empty_series_if_keyerror
-        def func(n, c):
-            col = n.df(c)[f"{nominal_attrs[c]}_opt"]
+        def func(n, c, port):
+            efficiency = abs(port_efficiency(n, c, port=port))
+            col = n.df(c)[f"{nominal_attrs[c]}_opt"] * efficiency
             if storage and (c == "StorageUnit"):
                 col = col * n.df(c).max_hours
-            if bus_carrier is None:
-                return col
-            else:
-                sign = n.df(c).get("sign", 1)
-                ports = get_ports(n, c)
-                ds = pd.Series(0, index=n.df(c).index)
-                for port in ports:
-                    mask = port_efficiency(n, c, port=port, bus_carrier=bus_carrier)
-                    df = sign * mask * col
-                    df.clip(lower=0, inplace=True)
-                    ds = ds.add(df, fill_value=0.0)
-                return ds[ds != 0]
+            return col
 
         df = aggregate_components(
             n,
@@ -511,6 +576,8 @@ class StatisticsAccessor:
             comps=comps,
             agg=aggregate_groups,
             groupby=groupby,
+            at_port=at_port,
+            bus_carrier=bus_carrier,
             nice_names=nice_names,
         )
         df.attrs["name"] = "Optimal Capacity"
@@ -522,6 +589,7 @@ class StatisticsAccessor:
         comps=None,
         aggregate_groups="sum",
         groupby=None,
+        at_port=False,
         bus_carrier=None,
         storage=False,
         nice_names=True,
@@ -529,7 +597,7 @@ class StatisticsAccessor:
         """
         Calculate the installed capacity of the network components in MW.
 
-        If `bus_carrier` is given, the capacity is weighed by the output efficiency
+        If `at_port` is True, the capacity is weighed by the output efficiency
         of components at buses with carrier `bus_carrier`.
 
         If storage is set to True, only storage capacities of the component
@@ -544,22 +612,12 @@ class StatisticsAccessor:
             comps = ("Store", "StorageUnit")
 
         @pass_empty_series_if_keyerror
-        def func(n, c):
-            col = n.df(c)[f"{nominal_attrs[c]}"]
+        def func(n, c, port):
+            efficiency = abs(port_efficiency(n, c, port=port))
+            col = n.df(c)[f"{nominal_attrs[c]}"] * efficiency
             if storage and (c == "StorageUnit"):
                 col = col * n.df(c).max_hours
-            if bus_carrier is None:
-                return col
-            else:
-                sign = n.df(c).get("sign", 1)
-                ports = get_ports(n, c)
-                ds = pd.Series(0, index=n.df(c).index)
-                for port in ports:
-                    mask = port_efficiency(n, c, port=port, bus_carrier=bus_carrier)
-                    df = sign * mask * col
-                    df.clip(lower=0, inplace=True)
-                    ds = ds.add(df, fill_value=0.0)
-                return ds[ds != 0]
+            return col
 
         df = aggregate_components(
             n,
@@ -567,6 +625,8 @@ class StatisticsAccessor:
             comps=comps,
             agg=aggregate_groups,
             groupby=groupby,
+            at_port=at_port,
+            bus_carrier=bus_carrier,
             nice_names=nice_names,
         )
         df.attrs["name"] = "Installed Capacity"
@@ -578,6 +638,7 @@ class StatisticsAccessor:
         comps=None,
         aggregate_groups="sum",
         groupby=None,
+        at_port=False,
         bus_carrier=None,
         nice_names=True,
     ):
@@ -594,15 +655,17 @@ class StatisticsAccessor:
             comps=comps,
             aggregate_groups=aggregate_groups,
             groupby=groupby,
-            nice_names=nice_names,
+            at_port=at_port,
             bus_carrier=bus_carrier,
+            nice_names=nice_names,
         ).sub(
             self.installed_capacity(
                 comps=comps,
                 aggregate_groups=aggregate_groups,
                 groupby=groupby,
-                nice_names=nice_names,
+                at_port=at_port,
                 bus_carrier=bus_carrier,
+                nice_names=nice_names,
             ),
             fill_value=0,
         )
@@ -616,6 +679,7 @@ class StatisticsAccessor:
         aggregate_time="sum",
         aggregate_groups="sum",
         groupby=None,
+        at_port=False,
         bus_carrier=None,
         nice_names=True,
     ):
@@ -638,23 +702,15 @@ class StatisticsAccessor:
         n = self._parent
 
         @pass_empty_series_if_keyerror
-        def func(n, c):
+        def func(n, c, port):
             if c in n.branch_components:
                 p = n.pnl(c).p0
             elif c == "StorageUnit":
                 p = n.pnl(c).p_dispatch
             else:
                 p = n.pnl(c).p
-            if bus_carrier is not None:
-                masks = [
-                    port_mask(n, c, port=port, bus_carrier=bus_carrier)
-                    for port in get_ports(n, c)
-                ]
-                mask = reduce(np.logical_or, masks)
-                p = p * mask
 
             opex = p * n.get_switchable_as_dense(c, "marginal_cost")
-            opex = opex.loc[:, (opex != 0).any()]
             weights = get_weightings(n, c)
             return aggregate_timeseries(opex, weights, agg=aggregate_time)
 
@@ -664,6 +720,8 @@ class StatisticsAccessor:
             comps=comps,
             agg=aggregate_groups,
             groupby=groupby,
+            at_port=at_port,
+            bus_carrier=bus_carrier,
             nice_names=nice_names,
         )
         df.attrs["name"] = "Operational Expenditure"
@@ -676,6 +734,7 @@ class StatisticsAccessor:
         aggregate_time="sum",
         aggregate_groups="sum",
         groupby=None,
+        at_port=True,
         bus_carrier=None,
         nice_names=True,
     ):
@@ -691,29 +750,15 @@ class StatisticsAccessor:
         """
         n = self._parent
 
-        @pass_empty_series_if_keyerror
-        def func(n, c):
-            sign = -1.0 if c in n.branch_components else n.df(c).get("sign", 1.0)
-            ports = get_ports(n, c)
-            p = pd.DataFrame(0, index=n.snapshots, columns=n.df(c).index)
-            for port in ports:
-                mask = port_mask(n, c, port=port, bus_carrier=bus_carrier)
-                df = n.pnl(c)[f"p{port}"]
-                df = sign * mask * df.reindex(columns=p.columns, fill_value=0)
-                df.clip(lower=0, inplace=True)
-                p = p.add(df, fill_value=0.0)
-            p = p.loc[:, (p != 0).any()]
-
-            weights = get_weightings(n, c)
-            return aggregate_timeseries(p, weights, agg=aggregate_time)
-
-        df = aggregate_components(
-            n,
-            func,
+        df = self.energy_balance(
             comps=comps,
-            agg=aggregate_groups,
+            aggregate_time=aggregate_time,
+            aggregate_groups=aggregate_groups,
             groupby=groupby,
+            at_port=at_port,
+            bus_carrier=bus_carrier,
             nice_names=nice_names,
+            kind="supply",
         )
         df.attrs["name"] = "Supply"
         df.attrs["unit"] = "carrier dependent"
@@ -725,6 +770,7 @@ class StatisticsAccessor:
         aggregate_time="sum",
         aggregate_groups="sum",
         groupby=None,
+        at_port=True,
         bus_carrier=None,
         nice_names=True,
     ):
@@ -740,42 +786,34 @@ class StatisticsAccessor:
         """
         n = self._parent
 
-        @pass_empty_series_if_keyerror
-        def func(n, c):
-            sign = -1.0 if c in n.branch_components else n.df(c).get("sign", 1.0)
-            ports = get_ports(n, c)
-            p = pd.DataFrame(0, index=n.snapshots, columns=n.df(c).index)
-            for port in ports:
-                mask = port_mask(n, c, port=port, bus_carrier=bus_carrier)
-                df = n.pnl(c)[f"p{port}"]
-                df = sign * mask * df.reindex(columns=p.columns, fill_value=0)
-                df = -df.clip(upper=0)
-                p = p.add(df, fill_value=0.0)
-            p = p.loc[:, (p != 0).any()]
-
-            weights = get_weightings(n, c)
-            return aggregate_timeseries(p, weights, agg=aggregate_time)
-
-        df = aggregate_components(
-            n,
-            func,
+        df = self.energy_balance(
             comps=comps,
-            agg=aggregate_groups,
+            aggregate_time=aggregate_time,
+            aggregate_groups=aggregate_groups,
             groupby=groupby,
+            at_port=at_port,
+            bus_carrier=bus_carrier,
             nice_names=nice_names,
+            kind="withdrawal",
         )
         df.attrs["name"] = "Withdrawal"
         df.attrs["unit"] = "carrier dependent"
         return df
 
+    # TODO: remove in 0.29
+    @deprecated(
+        deprecated_in="0.28", removed_in="0.29", details="Use 'energy_balance' instead."
+    )
     def dispatch(
         self,
         comps=None,
         aggregate_time="sum",
         aggregate_groups="sum",
         groupby=None,
+        at_port=True,
         bus_carrier=None,
         nice_names=True,
+        kind=None,
     ):
         """
         Calculate the dispatch of components in the network. Units depend on
@@ -794,30 +832,18 @@ class StatisticsAccessor:
             Note that for {'mean', 'sum'} the time series are aggregated to MWh
             using snapshot weightings. With False the time series is given in MW. Defaults to 'sum'.
         """
+
         n = self._parent
 
-        @pass_empty_series_if_keyerror
-        def func(n, c):
-            sign = -1.0 if c in n.branch_components else n.df(c).get("sign", 1.0)
-            ports = get_ports(n, c)
-            p = pd.DataFrame(0, index=n.snapshots, columns=n.df(c).index)
-            for port in ports:
-                mask = port_mask(n, c, port=port, bus_carrier=bus_carrier)
-                df = n.pnl(c)[f"p{port}"]
-                df = sign * mask * df.reindex(columns=p.columns, fill_value=0)
-                p = p.add(df, fill_value=0.0)
-            p = p.loc[:, (p != 0).any()]
-
-            weights = get_weightings(n, c)
-            return aggregate_timeseries(p, weights, agg=aggregate_time)
-
-        df = aggregate_components(
-            n,
-            func,
+        df = self.energy_balance(
             comps=comps,
-            agg=aggregate_groups,
+            aggregate_time=aggregate_time,
+            aggregate_groups=aggregate_groups,
             groupby=groupby,
+            at_port=at_port,
+            bus_carrier=bus_carrier,
             nice_names=nice_names,
+            kind=kind,
         )
         df.attrs["name"] = "Dispatch"
         df.attrs["unit"] = "carrier dependent"
@@ -829,6 +855,7 @@ class StatisticsAccessor:
         aggregate_time="sum",
         aggregate_groups="sum",
         groupby=None,
+        at_port=False,
         bus_carrier=None,
         nice_names=True,
     ):
@@ -856,8 +883,8 @@ class StatisticsAccessor:
         transmission_branches = get_transmission_branches(n, bus_carrier)
 
         @pass_empty_series_if_keyerror
-        def func(n, c):
-            p = n.pnl(c).p0[transmission_branches.get_loc_level(c)[1]]
+        def func(n, c, port):
+            p = n.pnl(c)[f"p{port}"][transmission_branches.get_loc_level(c)[1]]
             weights = get_weightings(n, c)
             return aggregate_timeseries(p, weights, agg=aggregate_time)
 
@@ -867,6 +894,8 @@ class StatisticsAccessor:
             comps=comps,
             agg=aggregate_groups,
             groupby=groupby,
+            at_port=at_port,
+            bus_carrier=bus_carrier,
             nice_names=nice_names,
         )
         df.attrs["name"] = "Transmission"
@@ -879,7 +908,11 @@ class StatisticsAccessor:
         aggregate_time="sum",
         aggregate_groups="sum",
         aggregate_bus=True,
+        groupby=get_carrier_and_bus_carrier,
+        at_port=True,
+        bus_carrier=None,
         nice_names=True,
+        kind=None,
     ):
         """
         Calculate the energy balance of components in the network. Positive
@@ -898,38 +931,60 @@ class StatisticsAccessor:
             Note that for {'mean', 'sum'} the time series are aggregated to MWh
             using snapshot weightings. With False the time series is given in MW. Defaults to 'sum'.
         """
+
         n = self._parent
 
-        @pass_empty_series_if_keyerror
-        def func(n, c):
-            sign = -1 if c in n.branch_components else n.df(c).get("sign", 1)
-            ports = [col[3:] for col in n.df(c).columns if col[:3] == "bus"]
-            p = []
-            for port in ports:
-                mask = n.df(c)[f"bus{port}"] != ""
-                df = sign * n.pnl(c)[f"p{port}"].loc[:, mask]
-                index = get_bus_and_carrier_and_bus_carrier(
-                    n, c, port=port, nice_names=nice_names
-                )
-                index = pd.concat(index, axis=1)[mask]
-                df.columns = pd.MultiIndex.from_frame(index.reindex(df.columns))
-                p.append(df)
-            p = pd.concat(p, axis=1)
-            weights = get_weightings(n, c)
-            return aggregate_timeseries(p, weights, agg=aggregate_time)
+        if (
+            n.buses.carrier.unique().size > 1
+            and groupby is None
+            and bus_carrier is None
+        ):
+            logger.warning(
+                "Network has multiple bus carriers which are aggregated together. To separate bus carriers set `bus_carrier` or use groupers like `get_carrier_and_bus_carrier`."
+            )
 
-        groupby = ["carrier", "bus_carrier"]
+        # TODO remove aggregate_bus in 0.29
         if not aggregate_bus:
-            groupby.append("bus")
+            if groupby != get_carrier_and_bus_carrier:
+                logger.warning(
+                    "Argument 'groupby' is ignored when 'aggregate_bus' is set to True. Falling back to default."
+                )
+            logger.warning(
+                "Argument 'aggregate_bus' is deprecated in 0.28 and will be removed in 0.29. Use grouper `get_bus_and_carrier_and_bus_carrier` instead."
+            )
+            switch = True
+            groupby = get_bus_and_carrier_and_bus_carrier
+        else:
+            switch = False
+
+        @pass_empty_series_if_keyerror
+        def func(n, c, port):
+            sign = -1.0 if c in n.branch_components else n.df(c).get("sign", 1.0)
+            weights = get_weightings(n, c)
+            p = sign * n.pnl(c)[f"p{port}"]
+            if kind == "supply":
+                p = p.clip(lower=0)
+            elif kind == "withdrawal":
+                p = -p.clip(upper=0)
+            elif kind != None:
+                logger.warning(
+                    f"Argument 'kind' is not recognized. Falling back to energy balance."
+                )
+            return aggregate_timeseries(p, weights, agg=aggregate_time)
 
         df = aggregate_components(
             n,
             func,
             comps=comps,
             agg=aggregate_groups,
-            groupby={"level": groupby},
+            groupby=groupby,
+            at_port=at_port,
+            bus_carrier=bus_carrier,
             nice_names=nice_names,
         )
+
+        if switch:
+            res = df.reorder_levels(["component", "carrier", "bus_carrier", "bus"])
         df.attrs["name"] = "Energy Balance"
         df.attrs["unit"] = "carrier dependent"
         return df
@@ -940,6 +995,7 @@ class StatisticsAccessor:
         aggregate_time="sum",
         aggregate_groups="sum",
         groupby=None,
+        at_port=False,
         bus_carrier=None,
         nice_names=True,
     ):
@@ -965,16 +1021,8 @@ class StatisticsAccessor:
         n = self._parent
 
         @pass_empty_series_if_keyerror
-        def func(n, c):
+        def func(n, c, port):
             p = (n.pnl(c).p_max_pu * n.df(c).p_nom_opt - n.pnl(c).p).clip(lower=0)
-            if bus_carrier is not None:
-                masks = [
-                    port_mask(n, c, port=port, bus_carrier=bus_carrier)
-                    for port in get_ports(n, c)
-                ]
-                mask = reduce(np.logical_or, masks)
-                p = p.loc[:, mask.astype(bool)]
-
             weights = get_weightings(n, c)
             return aggregate_timeseries(p, weights, agg=aggregate_time)
 
@@ -984,6 +1032,8 @@ class StatisticsAccessor:
             comps=comps,
             agg=aggregate_groups,
             groupby=groupby,
+            at_port=at_port,
+            bus_carrier=bus_carrier,
             nice_names=nice_names,
         )
         df.attrs["name"] = "Curtailment"
@@ -995,6 +1045,7 @@ class StatisticsAccessor:
         comps=None,
         aggregate_time="mean",
         aggregate_groups="sum",
+        at_port=False,
         groupby=None,
         bus_carrier=None,
         nice_names=True,
@@ -1018,30 +1069,20 @@ class StatisticsAccessor:
         n = self._parent
 
         @pass_empty_series_if_keyerror
-        def func(n, c):
+        def func(n, c, port):
             p = get_operation(n, c).abs()
-            if bus_carrier is not None:
-                masks = [
-                    port_mask(n, c, port=port, bus_carrier=bus_carrier)
-                    for port in get_ports(n, c)
-                ]
-                mask = reduce(np.logical_or, masks)
-                p = p.loc[:, mask.astype(bool)]
-
             weights = get_weightings(n, c)
             return aggregate_timeseries(p, weights, agg=aggregate_time)
 
-        df = aggregate_components(
-            n,
-            func,
+        kwargs = dict(
             comps=comps,
-            agg=aggregate_groups,
             groupby=groupby,
+            at_port=at_port,
+            bus_carrier=bus_carrier,
             nice_names=nice_names,
         )
-        capacity = self.optimal_capacity(
-            comps=comps, aggregate_groups=aggregate_groups, groupby=groupby
-        )
+        df = aggregate_components(n, func, agg=aggregate_groups, **kwargs)
+        capacity = self.optimal_capacity(aggregate_groups=aggregate_groups, **kwargs)
         df = df.div(capacity.reindex(df.index), axis=0)
         df.attrs["name"] = "Capacity Factor"
         df.attrs["unit"] = "p.u."
@@ -1053,14 +1094,19 @@ class StatisticsAccessor:
         aggregate_time="sum",
         aggregate_groups="sum",
         groupby=None,
+        at_port=True,
         bus_carrier=None,
         nice_names=True,
+        kind=None,
     ):
         """
         Calculate the revenue of components in the network in given currency.
+        The revenue is defined as the net revenue of an asset, i.e cost of input - revenue of output.
+        If kind is set to "input" or "output" only the revenue of the input or output is considered.
 
         If `bus_carrier` is given, only the revenue resulting from buses with carrier
         `bus_carrier` is considered.
+
 
         For information on the list of arguments, see the docs in
         `Network.statistics` or `pypsa.statistics.StatisticsAccessor`.
@@ -1071,26 +1117,30 @@ class StatisticsAccessor:
             Type of aggregation when aggregating time series.
             Note that for {'mean', 'sum'} the time series are aggregated to
             using snapshot weightings. With False the time series is given. Defaults to 'sum'.
+        kind : str, optional
+            Type of revenue to consider. If 'input' only the revenue of the input is considered.
+            If 'output' only the revenue of the output is considered. Defaults to None.
         """
         n = self._parent
 
         @pass_empty_series_if_keyerror
-        def func(n, c):
+        def func(n, c, port):
             sign = -1.0 if c in n.branch_components else n.df(c).get("sign", 1.0)
-            ports = get_ports(n, c)
-            revenue = pd.DataFrame(0, index=n.snapshots, columns=n.df(c).index)
-            for port in ports:
-                mask = port_mask(n, c, port=port, bus_carrier=bus_carrier)
-                df = n.pnl(c)[f"p{port}"]
-                df = sign * mask * df.reindex(columns=revenue.columns, fill_value=0)
-                buses = n.df(c)[f"bus{port}"][df.columns]
-                prices = n.buses_t.marginal_price.reindex(
-                    columns=buses, fill_value=0
-                ).values
-                df = df * prices
-                revenue = revenue.add(df, fill_value=0.0)
-
-            revenue = revenue.loc[:, (revenue != 0).any()]
+            df = sign * n.pnl(c)[f"p{port}"]
+            buses = n.df(c)[f"bus{port}"][df.columns]
+            prices = n.buses_t.marginal_price.reindex(
+                columns=buses, fill_value=0
+            ).values
+            if kind is not None:
+                if kind == "input":
+                    df = df.clip(upper=0)
+                elif kind == "output":
+                    df = df.clip(lower=0)
+                else:
+                    raise ValueError(
+                        f"Argument 'kind' must be 'input', 'output' or None, got {kind}"
+                    )
+            revenue = df * prices
             weights = get_weightings(n, c)
             return aggregate_timeseries(revenue, weights, agg=aggregate_time)
 
@@ -1100,6 +1150,8 @@ class StatisticsAccessor:
             comps=comps,
             agg=aggregate_groups,
             groupby=groupby,
+            at_port=at_port,
+            bus_carrier=bus_carrier,
             nice_names=nice_names,
         )
         df.attrs["name"] = "Revenue"
@@ -1112,6 +1164,7 @@ class StatisticsAccessor:
         aggregate_time="mean",
         aggregate_groups="sum",
         groupby=None,
+        at_port=True,
         bus_carrier=None,
         nice_names=True,
     ):
@@ -1138,8 +1191,9 @@ class StatisticsAccessor:
             aggregate_time=aggregate_time,
             aggregate_groups=aggregate_groups,
             groupby=groupby,
-            nice_names=nice_names,
+            at_port=at_port,
             bus_carrier=bus_carrier,
+            nice_names=nice_names,
         )
         df = self.revenue(**kwargs) / self.supply(**kwargs)
         df.attrs["name"] = "Market Value"